Update build script and add css-minimizer-webpack-plugin dependency
This commit updates the build command in the package.json file to specify the production webpack configuration file.
The css-minimizer-webpack-plugin has also been added to our list of devDependencies.
This is done to improve the minification of our CSS during the production build process and ensure that the right webpack configuration file is used when building for production.

The package-lock.json has also been regenerated due to the new dependency.