Refactor Character class and its subclasses
Simplified the Character class and subclasses by removing the generic type. The change makes the class more straightforward and less error-prone since "generic" was a prohibited value for the type of Character, and removing the generic types means that we can no longer attempt to create a character of a non-existent type. To prevent creating instances of Character, a TypeError is thrown if done so, making sure characters of specific types are created through the subclasses. Also, this refactor results in less code repetition.