Update Jest dependencies and stop excluding Jest test files
Jest dependencies in the 'package.json' and 'package-lock.json' were updated to address bugfixes and performance improvements in the newer versions. The 'tsconfig.json' file was also updated; the specific exclusion of the "__tests__" subdirectory in the source directory has been removed and is now applied to the entire "src" directory. This was necessary to enable the correct typescript compilation of Jest test files.