Add findSelectedCell method and enable outDir in tsconfig
A findSelectedCell method has been added to the GamePlay class in the 'src/ts/GamePlay.ts' file. This method returns the index of the first cell in the array that has the class "selected".

The outDir option has been enabled in the tsconfig.json file. This option allows the compiler to redirect the output structure to the 'dist' directory, improving the project's organization.