Refactor game character movement code for clarity and functionality
The game character's movement logic was restructured in GameController.ts to improve code readability and to ensure that a character only moves when a valid position is selected. This also takes into account that the move only happens when the action is valid for the character type. In addition, 'deselectAllCells' function in GamePlay.ts was added to deselect all cells after the character movement, improving UX.