Add game state to track turns in gameplay
Added a private gameState object in the GameController.ts. The GameState.ts class now has a 'turn' property to keep track of whose turn it is, enhancing synchronicity in the game. This gameState object gets initialized when the GameController is constructing. A static method has also been upgraded to take a GameState object and rotate turns. This will help ensure turn-based functionality works as expected.