Implement computer game logic and refactor character movement
This commit features a key addition to the game logic that allows computer-controlled characters to move logically and attack when necessary. The changes introduced seek to provide a more interactive game scenario by stimulating a sense-like enemy response during the computer's turn. With the aid of the 'getEuclideanDistance' and 'moveTowardsEnemy' methods, the enemy characters now have the ability to tactically calculate the distance between themselves and the player's characters, and to make movements that suggest an intelligent approach to the gameplay. These changes greatly enhance the player engagement since it now feels more like playing against another player.

In addition to the game logic adjustment, notable changes were made to the movement code of the character, aimed at enhancing code organization and improving readability. These changes include the conversion of the 'onCellClick' method to an asynchronous one for handling character damages and a conditional statement for passing the turn to the computer if it is indeed the computer's turn. These changes contribute to a more efficient and user-friendly gaming experience.