Refactor possible pass calculation and remove redundancy
Refactored the calculation of possible actions by directly invoking the 'possiblePassArrayGenerator' function once in the 'checkPass' method. Also, the redundant 'possibleActionArrayCleaner' method is removed. This was possible because the logic of filtering was moved inside the 'possiblePassArrayGenerator'. Furthermore, the calculation of Euclidean distance moved to utils.ts from GameController.ts. This makes the function reusable, improving code maintainability and readability.