Refactor level up logic and handle game over state
Updated the level up method in Character.ts to accept level as a parameter. This allows for more controlled leveling up outside the usual increments. A game over state check has been added to the GameController class. The GameController checks if the player's team has been wiped out or if the game reaches level 4, in which case the game ends and relevant actions are taken (i.e., deselecting all cells, resetting cursor and clearing board). This allows for better observance of game rules and improved user experience. The gameOver method has been introduced to handle all actions required when the game ends.