Enhance button transitions in style.css
Formatting in the button transitions of style.css has been improved for readability. Additional button states for "hover" and "active" have also been added. The "hover" state provides visual feedback to users when hovering over the buttons, and "active" changes the background-color to 'dimgrey' when the button is pressed. The changes enhance user interaction and overall user experience.