Optimize character deselection post death
Moved the 'deselectAllCells' method from the damage calculation portion, to the 'deathFitrator' method. This is done to ensure that characters are deselected only when a character is actually removed from the game State due to its health dropping to zero or below. Added a 'turn to computer' in the 'newLevel' method for better turn management. Ensures proper turn cycles and overall game structure.