Add validation for cell selections in game
Extra validations have been added in 'GameController.ts' to enhance gameplay. An if statement has been added to the 'onCellClick' event to check if the clicked cell is part of the game board, which avoids out of bounds issues. Additionally, the new methods 'checkSelectedEmptyCell' and 'checkSelectedemptyEnemyCell' are added to ensure that players are only selecting valid game board cells. These changes aim to prevent game play errors and enhance the user experience.