Update .appveyor.yml for GitHub deployment
Updated .appveyor.yml for deploying to GitHub Pages. Includes environment variables for securing the GitHub token and changing the git branch from 'master' to 'main'. Added git configs and deploy_script to push the build files to the 'gh-pages' branch. This enhances the CI/CD pipeline and allows automatic build and deployment on AppVeyor.